chore(EMS-3237): no pdf - change your answers - export contract - split up tests
