+++ .github/workflows/test.yml
@@ -203,6 +203,12 @@ jobs:
             "application-submission/single-policy-type/**/*.spec.js",
             "cannot-skip-flow/**/*.spec.js",
             "check-your-answers/export-contract/**/*.spec.js",
+            "check-your-answers/export-contract/about-goods-or-services/*.spec.js",
+            "check-your-answers/export-contract/agent/*.spec.js",
+            "check-your-answers/export-contract/agent-charges/*.spec.js",
+            "check-your-answers/export-contract/agent-service/*.spec.js",
+            "check-your-answers/export-contract/how-will-you-get-paid/*.spec.js",
+            "check-your-answers/export-contract/private-market/*.spec.js",
             "check-your-answers/policy/*.spec.js",
             "check-your-answers/policy/change-your-answers/broker/*.spec.js",
             "check-your-answers/policy/change-your-answers/multiple-policy/*.spec.js",
