+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/export-contract/change-your-answers/agent-service/change-your-answers-agent-service-charging-no-to-yes.spec.js
@@ -1,7 +1,8 @@
-import { status, summaryList } from '../../../../../../../pages/shared';
-import partials from '../../../../../../../partials';
-import FIELD_IDS from '../../../../../../../constants/field-ids/insurance/export-contract';
-import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
+import { status, summaryList } from '../../../../../../../../pages/shared';
+import partials from '../../../../../../../../partials';
+import FIELD_IDS from '../../../../../../../../constants/field-ids/insurance/export-contract';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+import checkSummaryList from '../../../../../../../../commands/insurance/check-export-contract-summary-list';
 
 const {
   ROOT,
@@ -10,11 +11,15 @@ const {
   },
   EXPORT_CONTRACT: {
     AGENT_SERVICE_CHECK_AND_CHANGE,
+    AGENT_CHARGES_CHECK_AND_CHANGE,
   },
 } = INSURANCE_ROUTES;
 
 const {
-  AGENT_SERVICE: { SERVICE_DESCRIPTION: FIELD_ID },
+  AGENT_SERVICE: { IS_CHARGING: FIELD_ID },
+  AGENT_CHARGES: {
+    FIXED_SUM_AMOUNT, PERCENTAGE_CHARGE, PAYABLE_COUNTRY_CODE,
+  },
 } = FIELD_IDS;
 
 const { taskList } = partials.insurancePartials;
@@ -23,9 +28,10 @@ const task = taskList.submitApplication.tasks.checkAnswers;
 
 const baseUrl = Cypress.config('baseUrl');
 
-context('Insurance - Change your answers - Export contract - Summary list - Agent service - description', () => {
+context('Insurance - Change your answers - Export contract - Summary list - Agent service - not charging - No to yes', () => {
   let referenceNumber;
   let url;
+  let agentChargesUrl;
 
   before(() => {
     cy.completeSignInAndGoToApplication({}).then(({ referenceNumber: refNumber }) => {
@@ -43,6 +49,7 @@ context('Insurance - Change your answers - Export contract - Summary list - Agen
       cy.completeAndSubmitMultipleCheckYourAnswers({ count: 3 });
 
       url = `${baseUrl}${ROOT}/${referenceNumber}${EXPORT_CONTRACT}`;
+      agentChargesUrl = `${baseUrl}${ROOT}/${referenceNumber}${AGENT_CHARGES_CHECK_AND_CHANGE}`;
 
       cy.assertUrl(url);
     });
@@ -69,27 +76,40 @@ context('Insurance - Change your answers - Export contract - Summary list - Agen
       });
     });
 
-    describe('form submission with a new answer', () => {
-      const newAnswer = 'Mock new agent service description';
-
+    describe('after changing the answer from no to yes', () => {
       beforeEach(() => {
+        cy.navigateToUrl(url);
+      });
+
+      it(`should redirect to ${EXPORT_CONTRACT} after completing (now required) ${AGENT_CHARGES_CHECK_AND_CHANGE} fields`, () => {
+        summaryList.field(FIELD_ID).changeLink().click();
+
         cy.navigateToUrl(url);
 
         summaryList.field(FIELD_ID).changeLink().click();
 
         cy.completeAndSubmitAgentServiceForm({
-          serviceDescription: newAnswer,
-          agentIsCharging: false,
+          agentIsCharging: true,
         });
-      });
 
-      it(`should redirect to ${EXPORT_CONTRACT}`, () => {
+        const expectedUrl = `${agentChargesUrl}#${FIELD_ID}-label`;
+        cy.assertUrl(expectedUrl);
+
+        cy.completeAndSubmitAgentChargesForm({
+          fixedSumMethod: true,
+        });
+
         cy.assertChangeAnswersPageUrl({ referenceNumber, route: EXPORT_CONTRACT, fieldId: FIELD_ID });
       });
 
-      it('should render the new answer and retain a `completed` status tag', () => {
-        cy.assertSummaryListRowValue(summaryList, FIELD_ID, newAnswer);
+      it(`should render new ${FIELD_ID} answer and other agent charges fields`, () => {
+        checkSummaryList[FIELD_ID]({ shouldRender: true, isYes: true });
+        checkSummaryList[FIXED_SUM_AMOUNT]({ shouldRender: true });
+        checkSummaryList[PERCENTAGE_CHARGE]({ shouldRender: false });
+        checkSummaryList[PAYABLE_COUNTRY_CODE]({ shouldRender: true });
+      });
 
+      it('should retain a `completed` status tag', () => {
         cy.checkTaskStatusCompleted(status);
       });
     });
