+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/export-contract/change-your-answers/agent-service/change-your-answers-agent-service-description.spec.js
@@ -1,0 +1,97 @@
+import { status, summaryList } from '../../../../../../../../pages/shared';
+import partials from '../../../../../../../../partials';
+import FIELD_IDS from '../../../../../../../../constants/field-ids/insurance/export-contract';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+
+const {
+  ROOT,
+  CHECK_YOUR_ANSWERS: {
+    EXPORT_CONTRACT,
+  },
+  EXPORT_CONTRACT: {
+    AGENT_SERVICE_CHECK_AND_CHANGE,
+  },
+} = INSURANCE_ROUTES;
+
+const {
+  AGENT_SERVICE: { SERVICE_DESCRIPTION: FIELD_ID },
+} = FIELD_IDS;
+
+const { taskList } = partials.insurancePartials;
+
+const task = taskList.submitApplication.tasks.checkAnswers;
+
+const baseUrl = Cypress.config('baseUrl');
+
+context('Insurance - Change your answers - Export contract - Summary list - Agent service - description', () => {
+  let referenceNumber;
+  let url;
+
+  before(() => {
+    cy.completeSignInAndGoToApplication({}).then(({ referenceNumber: refNumber }) => {
+      referenceNumber = refNumber;
+
+      cy.completePrepareApplicationSinglePolicyType({
+        referenceNumber,
+        isUsingAgent: true,
+        agentIsCharging: false,
+      });
+
+      task.link().click();
+
+      // To get past previous "Check your answers" pages
+      cy.completeAndSubmitMultipleCheckYourAnswers({ count: 3 });
+
+      url = `${baseUrl}${ROOT}/${referenceNumber}${EXPORT_CONTRACT}`;
+
+      cy.assertUrl(url);
+    });
+  });
+
+  beforeEach(() => {
+    cy.saveSession();
+
+    cy.navigateToUrl(url);
+  });
+
+  after(() => {
+    cy.deleteApplication(referenceNumber);
+  });
+
+  describe(FIELD_ID, () => {
+    describe('when clicking the `change` link', () => {
+      it(`should redirect to ${AGENT_SERVICE_CHECK_AND_CHANGE}`, () => {
+        cy.navigateToUrl(url);
+
+        summaryList.field(FIELD_ID).changeLink().click();
+
+        cy.assertChangeAnswersPageUrl({ referenceNumber, route: AGENT_SERVICE_CHECK_AND_CHANGE, fieldId: FIELD_ID });
+      });
+    });
+
+    describe('form submission with a new answer', () => {
+      const newAnswer = 'Mock new agent service description';
+
+      beforeEach(() => {
+        cy.navigateToUrl(url);
+
+        summaryList.field(FIELD_ID).changeLink().click();
+
+        cy.completeAndSubmitAgentServiceForm({
+          serviceDescription: newAnswer,
+          agentIsCharging: false,
+        });
+      });
+
+      it(`should redirect to ${EXPORT_CONTRACT}`, () => {
+        cy.assertChangeAnswersPageUrl({ referenceNumber, route: EXPORT_CONTRACT, fieldId: FIELD_ID });
+      });
+
+      it('should render the new answer and retain a `completed` status tag', () => {
+        cy.assertSummaryListRowValue(summaryList, FIELD_ID, newAnswer);
+
+        cy.checkTaskStatusCompleted(status);
+      });
+    });
+  });
+});
